Put everything dangerous behind conditional comp
