+++ Robust.Client/Robust.Client.csproj
@@ -42,6 +42,9 @@
     <ProjectReference Include="..\Robust.LoaderApi\Robust.LoaderApi\Robust.LoaderApi.csproj" />
     <ProjectReference Include="..\Robust.Shared.Maths\Robust.Shared.Maths.csproj" />
     <ProjectReference Include="..\Robust.Shared\Robust.Shared.csproj" />
+  </ItemGroup>
+  
+  <ItemGroup Condition="'$(RobustToolsBuild)' == 'True'">
     <ProjectReference Include="..\RobustXaml\RobustXaml.csproj" />
   </ItemGroup>
