+++ Robust.Client/UserInterface/XAML/Proxy/IXamlHotReloadManager.cs
@@ -1,0 +1,6 @@
+﻿namespace Robust.Client.UserInterface.XAML.Proxy;
+
+interface IXamlHotReloadManager
+{
+    void Initialize();
+}
