+++ Robust.Client/UserInterface/XAML/Proxy/IXamlProxyManager.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Robust.Client.UserInterface.XAML.Proxy;
+
+public interface IXamlProxyManager
+{
+    void Initialize();
+    bool CanSetImplementation(string fileName);
+    void SetImplementation(string fileName, string fileContent);
+    bool Populate(Type t, object o);
+}
