+++ Robust.Client/UserInterface/XAML/Proxy/XamlHotReloadManager.Stub.cs
@@ -1,0 +1,22 @@
+﻿#if !TOOLS
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Robust.Shared.Asynchronous;
+using Robust.Shared.ContentPack;
+using Robust.Shared.IoC;
+using Robust.Shared.Log;
+
+namespace Robust.Client.UserInterface.XAML.Proxy
+{
+
+    internal sealed class XamlHotReloadManager : IXamlHotReloadManager
+    {
+        public void Initialize()
+        {
+            // This is a stub, so do nothing
+        }
+    }
+}
+#endif
