+++ Robust.Client/UserInterface/XAML/Proxy/XamlJitDelegate.Real.cs
@@ -1,6 +1,8 @@
-﻿using System;
+﻿#if TOOLS
+using System;
 using System.Reflection;
 
 namespace Robust.Client.UserInterface.XAML.Proxy;
 
 delegate MethodInfo? XamlJitDelegate(Type type, Uri uri, string filename, string content);
+#endif
